make_notificationをモジュール化  #6
