+++ src/app/arxiv/getPaperInfo.js
@@ -2,30 +2,7 @@ const http = require('http');
 const config = require('config');
 const xml_parser = require('./_xml_parser');
 const options2url = require('./_options2url');
-
-function make_notification(data_list) {
-  let result = '';
-  for (let i=0; i < data_list.length; i++) {
-    data = data_list[i];
-    result += `title:  ${data.title}\n`;
-    const authors = (authors_list) => {
-      let authors = '';
-      for (let j=0; j < authors_list.length; j++) {
-        authors += authors_list[i].name;
-        if (j != authors_list.length - 1) {
-          authors += ', ';
-        }
-      }
-      return authors;
-    };
-    result += `auhors: ${authors(data.author)}\n`;
-    result += `url: ${data.id}\n`;
-    if (i != data_list.length - 1) {
-      result += '\n ---------- \n';
-    }
-  }
-  return result;
-};
+const make_notification = require('./_make_notification');
 
 
 const get_paper_info = async () => {
